backend/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
pluggy==1.5.0
pytest==8.3.3
python-dotenv==1.0.1
virtualenv==20.26.4
Werkzeug==3.0.4
